- moving the code functions and states to useEditor state file.
